レイアウトの改善、コードの修正
