+++ app/src/main/java/com/saradabar/cpadcustomizetool/view/flagment/DeviceSettingsFragment.java
@@ -13,16 +13,10 @@
 package com.saradabar.cpadcustomizetool.view.flagment;
 
 import android.annotation.SuppressLint;
-import android.content.ActivityNotFoundException;
-import android.content.Intent;
-import android.content.pm.ApplicationInfo;
 import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
 import android.view.View;
-import android.widget.AbsListView;
-import android.widget.ListView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.widget.AppCompatEditText;
@@ -33,78 +27,29 @@ import androidx.preference.SwitchPreferenceCompat;
 
 import com.saradabar.cpadcustomizetool.R;
 import com.saradabar.cpadcustomizetool.util.Common;
-import com.saradabar.cpadcustomizetool.util.Constants;
 import com.saradabar.cpadcustomizetool.util.DialogUtil;
-import com.saradabar.cpadcustomizetool.view.activity.DeviceInfoActivity;
-import com.saradabar.cpadcustomizetool.view.activity.WebViewActivity;
-import com.saradabar.cpadcustomizetool.view.views.LaunchAppListView;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
-import java.util.List;
 import java.util.Locale;
 
-public class OtherFragment extends PreferenceFragmentCompat {
+public class DeviceSettingsFragment extends PreferenceFragmentCompat {
 
-    Preference preOtherStartSettings,
-            preStartUiAdjustment,
-            preStartDevSettings,
-            preScreenOffTimeOut,
-            preSleepTimeout,
-            preWebView,
-            preLaunchApp,
-            preDeviceInfo;
+    Preference preScreenOffTimeOut,
+            preSleepTimeout;
 
     SwitchPreferenceCompat swUiModeNight;
 
     @Override
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
-        addPreferencesFromResource(R.xml.pre_other);
+        setPreferencesFromResource(R.xml.pre_device_settings, rootKey);
 
-        preOtherStartSettings = findPreference("pre_other_start_settings");
-        preStartUiAdjustment = findPreference("pre_other_start_ui_adjustment");
-        preStartDevSettings = findPreference("pre_other_start_dev_settings");
         preScreenOffTimeOut = findPreference("pre_other_screen_off_time");
         preSleepTimeout = findPreference("pre_other_sleep_timeout");
-        preWebView = findPreference("pre_other_web_view");
-        preLaunchApp = findPreference("pre_other_launch_app");
-        preDeviceInfo = findPreference("pre_other_start_device_info");
         swUiModeNight = findPreference("pre_other_ui_mode_night");
-
-        preOtherStartSettings.setOnPreferenceClickListener(preference -> {
-            try {
-                startActivity(new Intent(Settings.ACTION_SETTINGS).addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION));
-            } catch (ActivityNotFoundException ignored) {
-            }
-            return false;
-        });
-
-        preStartDevSettings.setOnPreferenceClickListener(preference -> {
-            if (Settings.Secure.getInt(requireActivity().getContentResolver(), Settings.Global.DEVELOPMENT_SETTINGS_ENABLED, 0) == 1) {
-                try {
-                    // DchaCompletedPast && DchaState != 3
-                    if (Common.getDchaCompletedPast() && Settings.System.getInt(requireActivity().getContentResolver(), Constants.DCHA_STATE, 0) != 3) {
-                        Settings.System.putInt(requireActivity().getContentResolver(), Constants.DCHA_STATE, 3);
-                    }
-                    startActivity(new Intent(Settings.ACTION_APPLICATION_DEVELOPMENT_SETTINGS).addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION));
-                } catch (ActivityNotFoundException ignored) {
-                }
-            } else {
-                Toast.makeText(requireActivity(), R.string.toast_no_debug_option, Toast.LENGTH_SHORT).show();
-            }
-            return false;
-        });
-
-        preStartUiAdjustment.setOnPreferenceClickListener(preference -> {
-            try {
-                startActivity(new Intent().setClassName("com.android.systemui", "com.android.systemui.DemoMode").addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION));
-            } catch (ActivityNotFoundException ignored) {
-            }
-            return false;
-        });
 
         preScreenOffTimeOut.setOnPreferenceClickListener(preference -> {
             @SuppressLint("InflateParams") View view = requireActivity().getLayoutInflater().inflate(R.layout.view_time_out, null);
@@ -186,54 +131,6 @@ public class OtherFragment extends PreferenceFragmentCompat {
             return false;
         });
 
-        preWebView.setOnPreferenceClickListener(preference -> {
-            startActivity(new Intent(requireActivity(), WebViewActivity.class).putExtra("URL", "https://www.google.com/intl/ja/").addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION));
-            return false;
-        });
-
-        preLaunchApp.setOnPreferenceClickListener(preference -> {
-            @SuppressLint("InflateParams") View view = requireActivity().getLayoutInflater().inflate(R.layout.layout_launch_app_list, null);
-            List<ApplicationInfo> installedAppList = requireActivity().getPackageManager().getInstalledApplications(0);
-            List<LaunchAppListView.AppData> dataList = new ArrayList<>();
-
-            for (ApplicationInfo app : installedAppList) {
-                if(requireActivity().getPackageManager().getLaunchIntentForPackage(app.packageName) != null) {
-                    LaunchAppListView.AppData data = new LaunchAppListView.AppData();
-                    data.icon = app.loadIcon(requireActivity().getPackageManager());
-                    data.label = app.loadLabel(requireActivity().getPackageManager()).toString();
-                    data.packName = app.packageName;
-                    dataList.add(data);
-                }
-            }
-
-            ListView listView = view.findViewById(R.id.launch_app_list);
-            listView.setChoiceMode(AbsListView.CHOICE_MODE_SINGLE);
-            listView.setAdapter(new LaunchAppListView.LaunchAppAdapter(requireActivity(), dataList));
-            listView.setOnItemClickListener((parent, mView, position, id) -> {
-                try {
-                    startActivity(requireActivity().getPackageManager().getLaunchIntentForPackage(dataList.get(position).packName));
-                } catch (Exception e) {
-                    new DialogUtil(requireActivity())
-                            .setTitle(R.string.dialog_title_error)
-                            .setMessage(e.getMessage())
-                            .setPositiveButton(R.string.dialog_common_ok, null)
-                            .show();
-                }
-            });
-
-            new DialogUtil(requireActivity())
-                    .setView(view)
-                    .setTitle("アプリを選択")
-                    .setPositiveButton(R.string.dialog_common_cancel, null)
-                    .show();
-            return false;
-        });
-
-        preDeviceInfo.setOnPreferenceClickListener(preference -> {
-            startActivity(new Intent(requireActivity(), DeviceInfoActivity.class));
-            return false;
-        });
-
         swUiModeNight.setOnPreferenceChangeListener((preference, o) -> {
             if ((boolean) o) {
                 Common.exec("cmd uimode night yes");
@@ -244,15 +141,11 @@ public class OtherFragment extends PreferenceFragmentCompat {
             }
             return true;
         });
-        initPre();
+        initPreference();
     }
 
     /** @noinspection SequencedCollectionMethodCanBeUsed*/
-    private void initPre() {
-        if (Common.isCT2()) {
-            preStartUiAdjustment.setEnabled(false);
-            preStartUiAdjustment.setSummary(Build.MODEL + requireActivity().getString(R.string.pre_main_sum_message_1));
-        }
+    private void initPreference() {
         setSummaryScreenOffTimeConvert();
         ArrayList<String> stringArrayList = Common.exec("cmd uimode night");
 
@@ -264,6 +157,11 @@ public class OtherFragment extends PreferenceFragmentCompat {
                 swUiModeNight.setChecked(false);
                 swUiModeNight.setSummary("外観モードは、”ライトモード”に設定されています。");
             }
+        }
+
+        if (Common.isCT2()) {
+            swUiModeNight.setEnabled(false);
+            swUiModeNight.setSummary(Build.MODEL + getString(R.string.pre_main_sum_message_1));
         }
     }
